requirements.txt
langchain==0.3.7
langchain_anthropic==0.2.4
langchain_core==0.3.63
langchain_google_genai>=2.0.6
langchain_groq==0.2.1
pandas==2.2.3
python-dotenv==1.2.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.12
motor==3.3.2